+++ src/EMerchantPay/Resources/views/backend/emerchantpay_payment_methods/store/emerchantpay_wpf_tokenization_store.js
@@ -1,0 +1,18 @@
+// Emerchantpay WPF Tokenization
+
+
+
+
+Ext.define('Shopware.apps.EmerchantpayPaymentMethods.store.EmerchantpayWPFTokenizationStore', {
+    extend: 'Ext.data.Store',
+
+    fields: [
+        { name: 'option', type: 'string' },
+        { name: 'value', type: 'string' }
+    ],
+
+    data: [
+        { option: "{s name=emerchantpay/config/wpf_tonkenization_yes}Yes{/s}", value: 'yes' },
+        { option: "{s name=emerchantpay/config/wpf_tonkenization_no}No{/s}", value: 'no' }
+    ]
+});